moved github authorization into routes and added oauth2 middleware
