+++ server/routes/oauth2.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+
+const githubRoute = require('./github');
+
+//middleware routing
+router.use('/github', githubRoute);
+
+module.exports = router;
